load_balancer/requirements.txt
uvicorn
watchfiles

# psycopg2==2.8.6
fastapi==0.90.1

requests
httpx
pytest-cov
pytest-randomly
